requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.1
django-allauth==0.39.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
drf-writable-nested==0.5.1
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
PyJWT==1.7.1
PySocks==1.7.0
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
twilio==6.27.0
urllib3==1.24.3
